test: split app-middleware test per describe
