+++ test/e2e/app-dir/app-middleware/app-middleware.test.ts
@@ -1,8 +1,7 @@
 /* eslint-env jest */
-import path from 'path'
 import cheerio from 'cheerio'
 import { check, retry, withQuery } from 'next-test-utils'
-import { nextTestSetup, FileRef } from 'e2e-utils'
+import { nextTestSetup } from 'e2e-utils'
 import type { Response } from 'node-fetch'
 
 describe('app-dir with middleware', () => {
@@ -271,62 +270,4 @@ describe('app-dir with middleware', () => {
       )
     })
   }
-})
-
-describe('app dir - middleware without pages dir', () => {
-  const { next } = nextTestSetup({
-    files: {
-      app: new FileRef(path.join(__dirname, 'app')),
-      'next.config.js': new FileRef(path.join(__dirname, 'next.config.js')),
-      'middleware.js': `
-      import { NextResponse } from 'next/server'
-
-      export async function middleware(request) {
-        return new NextResponse('redirected')
-      }
-
-      export const config = {
-        matcher: '/headers'
-      }
-    `,
-    },
-  })
-
-  // eslint-disable-next-line jest/no-identical-title
-  it('Updates headers', async () => {
-    const html = await next.render('/headers')
-
-    expect(html).toContain('redirected')
-  })
-})
-
-describe('app dir - middleware with middleware in src dir', () => {
-  const { next } = nextTestSetup({
-    files: {
-      'src/app': new FileRef(path.join(__dirname, 'app')),
-      'next.config.js': new FileRef(path.join(__dirname, 'next.config.js')),
-      'src/middleware.js': `
-      import { NextResponse } from 'next/server'
-      import { cookies } from 'next/headers'
-
-      export async function middleware(request) {
-        const cookie = (await cookies()).get('test-cookie')
-        return NextResponse.json({ cookie })
-      }
-    `,
-    },
-  })
-
-  it('works without crashing when using RequestStore', async () => {
-    const browser = await next.browser('/')
-    await browser.addCookie({
-      name: 'test-cookie',
-      value: 'test-cookie-response',
-    })
-    await browser.refresh()
-
-    const html = await browser.eval('document.documentElement.innerHTML')
-
-    expect(html).toContain('test-cookie-response')
-  })
 })
